Rename artifact-uploader to artifactuploader to match golint rules
