+++ label_sync/labels.yaml
@@ -740,8 +740,8 @@ repos:
         target: both
         addedBy: label
       - color: 0052cc
-        description: Issues or PRs related to prow's artifact-uploader component
-        name: area/prow/artifact-uploader
+        description: Issues or PRs related to prow's artifactuploader component
+        name: area/prow/artifactuploader
         target: both
         addedBy: label
       - color: 0052cc
