+++ prow/artifactuploader/BUILD.bazel
@@ -3,7 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "go_default_library",
     srcs = ["controller.go"],
-    importpath = "k8s.io/test-infra/prow/artifact-uploader",
+    importpath = "k8s.io/test-infra/prow/artifactuploader",
     visibility = ["//visibility:public"],
     deps = [
         "//prow/gcsupload:go_default_library",
