+++ prow/cmd/artifactuploader/Dockerfile
@@ -16,4 +16,4 @@ FROM gcr.io/k8s-prow/alpine:0.1
 LABEL maintainer="skuznets@redhat.com"
 
 COPY artifact-uploader /artifact-uploader
-ENTRYPOINT ["/artifact-uploader"]
+ENTRYPOINT ["/artifactuploader"]
